pyproject.toml
[tool.poetry]
name = "fastapi-mailing-lists"
version = "0.1.0"
description = "An example of FastAPI backend with a mailing lists system"
authors = ["AliRzaev <alirzaev1996@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
uvicorn = "^0.17.0"
alembic = "^1.7.5"
SQLAlchemy = "^1.4.31"
pydantic = {extras = ["email"], version = "^1.9.0"}
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
emails = "^0.6"
asyncpg = "^0.25.0"
aioredis = "^2.0.1"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
